Refactor Map component into smaller components
